# artifacts/api-server/src/routes/schedules.ts
import { Router, type IRouter } from "express";
import { db } from "@workspace/db";
import {
  scheduleBlocksTable, staffTable, studentsTable, serviceTypesTable,
  staffAssignmentsTable, serviceRequirementsTable
} from "@workspace/db";
import {
  ListScheduleBlocksQueryParams,
  CreateScheduleBlockBody,
  UpdateScheduleBlockParams,
  UpdateScheduleBlockBody,
  DeleteScheduleBlockParams,
  GetScheduleConflictsQueryParams,
  GenerateScheduleBody,
  AcceptGeneratedScheduleBody,
  ListStaffAssignmentsQueryParams,
  CreateStaffAssignmentBody,
  DeleteStaffAssignmentParams,
} from "@workspace/api-zod";
import { eq, and } from "drizzle-orm";
import { computeAllActiveMinuteProgress } from "../lib/minuteCalc";

const router: IRouter = Router();

function blockToJson(b: any, staffName?: string | null, studentName?: string | null, serviceTypeName?: string | null) {
  return {
    id: b.id,
    staffId: b.staffId,
    staffName: staffName ?? null,
    studentId: b.studentId,
    studentName: studentName ?? null,
    serviceTypeId: b.serviceTypeId,
    serviceTypeName: serviceTypeName ?? null,
    dayOfWeek: b.dayOfWeek,
    startTime: b.startTime,
    endTime: b.endTime,
    location: b.location,
    blockLabel: b.blockLabel,
    blockType: b.blockType,
    notes: b.notes,
    isRecurring: b.isRecurring,
    weekOf: b.weekOf,
    isAutoGenerated: b.isAutoGenerated,
    createdAt: b.createdAt instanceof Date ? b.createdAt.toISOString() : b.createdAt,
  };
}

router.get("/schedule-blocks", async (req, res): Promise<void> => {
  const params = ListScheduleBlocksQueryParams.safeParse(req.query);
  const conditions: any[] = [];
  if (params.success) {
    if (params.data.staffId) conditions.push(eq(scheduleBlocksTable.staffId, Number(params.data.staffId)));
    if (params.data.studentId) conditions.push(eq(scheduleBlocksTable.studentId, Number(params.data.studentId)));
    if (params.data.dayOfWeek) conditions.push(eq(scheduleBlocksTable.dayOfWeek, params.data.dayOfWeek));
    if (params.data.weekOf) conditions.push(eq(scheduleBlocksTable.weekOf, params.data.weekOf));
  }

  const blocks = await db
    .select({
      id: scheduleBlocksTable.id,
      staffId: scheduleBlocksTable.staffId,
      studentId: scheduleBlocksTable.studentId,
      serviceTypeId: scheduleBlocksTable.serviceTypeId,
      dayOfWeek: scheduleBlocksTable.dayOfWeek,
      startTime: scheduleBlocksTable.startTime,
      endTime: scheduleBlocksTable.endTime,
      location: scheduleBlocksTable.location,
      blockLabel: scheduleBlocksTable.blockLabel,
      blockType: scheduleBlocksTable.blockType,
      notes: scheduleBlocksTable.notes,
      isRecurring: scheduleBlocksTable.isRecurring,
      weekOf: scheduleBlocksTable.weekOf,
      isAutoGenerated: scheduleBlocksTable.isAutoGenerated,
      createdAt: scheduleBlocksTable.createdAt,
      staffFirst: staffTable.firstName,
      staffLast: staffTable.lastName,
      studentFirst: studentsTable.firstName,
      studentLast: studentsTable.lastName,
      serviceTypeName: serviceTypesTable.name,
    })
    .from(scheduleBlocksTable)
    .leftJoin(staffTable, eq(staffTable.id, scheduleBlocksTable.staffId))
    .leftJoin(studentsTable, eq(studentsTable.id, scheduleBlocksTable.studentId))
    .leftJoin(serviceTypesTable, eq(serviceTypesTable.id, scheduleBlocksTable.serviceTypeId))
    .where(conditions.length > 0 ? and(...conditions) : undefined);

  res.json(blocks.map(b => blockToJson(b,
    b.staffFirst ? `${b.staffFirst} ${b.staffLast}` : null,
    b.studentFirst ? `${b.studentFirst} ${b.studentLast}` : null,
    b.serviceTypeName
  )));
});

router.post("/schedule-blocks", async (req, res): Promise<void> => {
  const parsed = CreateScheduleBlockBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [block] = await db.insert(scheduleBlocksTable).values({
    ...parsed.data,
    isAutoGenerated: false,
  }).returning();
  res.status(201).json(blockToJson(block));
});

router.patch("/schedule-blocks/:id", async (req, res): Promise<void> => {
  const params = UpdateScheduleBlockParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  const parsed = UpdateScheduleBlockBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const updateData: Partial<typeof scheduleBlocksTable.$inferInsert> = {};
  if (parsed.data.studentId !== undefined) updateData.studentId = parsed.data.studentId;
  if (parsed.data.startTime != null) updateData.startTime = parsed.data.startTime;
  if (parsed.data.endTime != null) updateData.endTime = parsed.data.endTime;
  if (parsed.data.location !== undefined) updateData.location = parsed.data.location;
  if (parsed.data.notes !== undefined) updateData.notes = parsed.data.notes;

  const [block] = await db.update(scheduleBlocksTable).set(updateData).where(eq(scheduleBlocksTable.id, params.data.id)).returning();
  if (!block) {
    res.status(404).json({ error: "Block not found" });
    return;
  }
  res.json(blockToJson(block));
});

router.delete("/schedule-blocks/:id", async (req, res): Promise<void> => {
  const params = DeleteScheduleBlockParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  await db.delete(scheduleBlocksTable).where(eq(scheduleBlocksTable.id, params.data.id));
  res.sendStatus(204);
});

// Conflicts detection
router.get("/schedule-blocks/conflicts", async (req, res): Promise<void> => {
  const allBlocks = await db
    .select({
      id: scheduleBlocksTable.id,
      staffId: scheduleBlocksTable.staffId,
      dayOfWeek: scheduleBlocksTable.dayOfWeek,
      startTime: scheduleBlocksTable.startTime,
      endTime: scheduleBlocksTable.endTime,
      studentId: scheduleBlocksTable.studentId,
      staffFirst: staffTable.firstName,
      staffLast: staffTable.lastName,
      studentFirst: studentsTable.firstName,
      studentLast: studentsTable.lastName,
    })
    .from(scheduleBlocksTable)
    .leftJoin(staffTable, eq(staffTable.id, scheduleBlocksTable.staffId))
    .leftJoin(studentsTable, eq(studentsTable.id, scheduleBlocksTable.studentId))
    .where(eq(scheduleBlocksTable.isRecurring, true));

  // Group by staff + day
  const grouped = new Map<string, typeof allBlocks>();
  for (const b of allBlocks) {
    const key = `${b.staffId}-${b.dayOfWeek}`;
    if (!grouped.has(key)) grouped.set(key, []);
    grouped.get(key)!.push(b);
  }

  const conflicts = [];
  for (const [_, blocks] of grouped.entries()) {
    if (blocks.length < 2) continue;
    const conflictingPairs: any[] = [];
    for (let i = 0; i < blocks.length; i++) {
      for (let j = i + 1; j < blocks.length; j++) {
        const a = blocks[i];
        const b = blocks[j];
        if (a.startTime < b.endTime && b.startTime < a.endTime) {
          // Calculate overlap
          const overlapStart = a.startTime > b.startTime ? a.startTime : b.startTime;
          const overlapEnd = a.endTime < b.endTime ? a.endTime : b.endTime;
          const [osH, osM] = overlapStart.split(":").map(Number);
          const [oeH, oeM] = overlapEnd.split(":").map(Number);
          const overlapMinutes = (oeH * 60 + oeM) - (osH * 60 + osM);

          conflictingPairs.push({
            staffId: a.staffId,
            staffName: a.staffFirst ? `${a.staffFirst} ${a.staffLast}` : `Staff ${a.staffId}`,
            dayOfWeek: a.dayOfWeek,
            conflictingBlocks: [
              { id: a.id, startTime: a.startTime, endTime: a.endTime, studentName: a.studentFirst ? `${a.studentFirst} ${a.studentLast}` : null },
              { id: b.id, startTime: b.startTime, endTime: b.endTime, studentName: b.studentFirst ? `${b.studentFirst} ${b.studentLast}` : null },
            ],
            overlapMinutes,
          });
        }
      }
    }
    conflicts.push(...conflictingPairs);
  }

  res.json(conflicts);
});

// Coverage gaps
router.get("/schedule-blocks/coverage-gaps", async (req, res): Promise<void> => {
  // Find students with active service requirements who have no scheduled blocks
  const reqs = await db
    .select({
      id: serviceRequirementsTable.id,
      studentId: serviceRequirementsTable.studentId,
      serviceTypeId: serviceRequirementsTable.serviceTypeId,
      requiredMinutes: serviceRequirementsTable.requiredMinutes,
      intervalType: serviceRequirementsTable.intervalType,
      studentFirst: studentsTable.firstName,
      studentLast: studentsTable.lastName,
      serviceTypeName: serviceTypesTable.name,
    })
    .from(serviceRequirementsTable)
    .leftJoin(studentsTable, eq(studentsTable.id, serviceRequirementsTable.studentId))
    .leftJoin(serviceTypesTable, eq(serviceTypesTable.id, serviceRequirementsTable.serviceTypeId))
    .where(eq(serviceRequirementsTable.active, true));

  const gaps = [];
  for (const req of reqs) {
    const blocks = await db
      .select({ id: scheduleBlocksTable.id })
      .from(scheduleBlocksTable)
      .where(and(
        eq(scheduleBlocksTable.studentId, req.studentId),
        eq(scheduleBlocksTable.serviceTypeId, req.serviceTypeId),
        eq(scheduleBlocksTable.isRecurring, true)
      ));

    if (blocks.length === 0) {
      gaps.push({
        studentId: req.studentId,
        studentName: req.studentFirst ? `${req.studentFirst} ${req.studentLast}` : `Student ${req.studentId}`,
        serviceRequirementId: req.id,
        serviceTypeName: req.serviceTypeName ?? "Unknown",
        gapDescription: `No scheduled blocks found for ${req.serviceTypeName ?? "this service"} (${req.requiredMinutes} min/${req.intervalType} required)`,
        severity: req.requiredMinutes > 60 ? "high" : "medium",
      });
    }
  }

  res.json(gaps);
});

// Auto-scheduler
router.post("/scheduler/generate", async (req, res): Promise<void> => {
  const parsed = GenerateScheduleBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { weekOf, staffIds, studentIds } = parsed.data;

  // Get all active service requirements
  const reqConditions: any[] = [eq(serviceRequirementsTable.active, true)];
  if (studentIds && studentIds.length > 0) {
    // Filter to specific students - simplified
  }

  const reqs = await db
    .select({
      id: serviceRequirementsTable.id,
      studentId: serviceRequirementsTable.studentId,
      providerId: serviceRequirementsTable.providerId,
      serviceTypeId: serviceRequirementsTable.serviceTypeId,
      requiredMinutes: serviceRequirementsTable.requiredMinutes,
      intervalType: serviceRequirementsTable.intervalType,
      studentFirst: studentsTable.firstName,
      studentLast: studentsTable.lastName,
      serviceTypeName: serviceTypesTable.name,
    })
    .from(serviceRequirementsTable)
    .leftJoin(studentsTable, eq(studentsTable.id, serviceRequirementsTable.studentId))
    .leftJoin(serviceTypesTable, eq(serviceTypesTable.id, serviceRequirementsTable.serviceTypeId))
    .where(eq(serviceRequirementsTable.active, true));

  const allStaff = await db.select().from(staffTable).where(eq(staffTable.status, "active"));

  const days = ["monday", "tuesday", "wednesday", "thursday", "friday"];
  const timeSlots = ["08:00", "09:00", "10:00", "11:00", "13:00", "14:00", "15:00"];
  const SLOT_DURATION = 60; // minutes

  const proposedBlocks: any[] = [];
  const unresolvedDeficits: any[] = [];

  // Track staff bookings to avoid double-booking
  const staffBookings = new Map<string, string[]>(); // key: staffId-day, value: array of "startTime-endTime"

  function isSlotAvailable(staffId: number, day: string, startTime: string): boolean {
    const key = `${staffId}-${day}`;
    const bookings = staffBookings.get(key) ?? [];
    const slotEnd = addMinutes(startTime, SLOT_DURATION);
    return !bookings.some(b => {
      const [bStart, bEnd] = b.split("-");
      return startTime < bEnd && slotEnd > bStart;
    });
  }

  function bookSlot(staffId: number, day: string, startTime: string) {
    const key = `${staffId}-${day}`;
    if (!staffBookings.has(key)) staffBookings.set(key, []);
    staffBookings.get(key)!.push(`${startTime}-${addMinutes(startTime, SLOT_DURATION)}`);
  }

  function addMinutes(time: string, minutes: number): string {
    const [h, m] = time.split(":").map(Number);
    const total = h * 60 + m + minutes;
    return `${String(Math.floor(total / 60)).padStart(2, "0")}:${String(total % 60).padStart(2, "0")}`;
  }

  // Simple constraint-based scheduler
  for (const req of reqs) {
    const provider = req.providerId ? allStaff.find(s => s.id === req.providerId) : null;
    if (!provider) {
      unresolvedDeficits.push({
        studentId: req.studentId,
        studentName: req.studentFirst ? `${req.studentFirst} ${req.studentLast}` : `Student ${req.studentId}`,
        serviceRequirementId: req.id,
        serviceTypeName: req.serviceTypeName ?? "Unknown",
        gapDescription: `No provider assigned for ${req.serviceTypeName}`,
        severity: "high",
      });
      continue;
    }

    // Determine sessions needed per week (rough estimate based on interval)
    let sessionsPerWeek = 1;
    if (req.intervalType === "weekly") sessionsPerWeek = Math.ceil(req.requiredMinutes / SLOT_DURATION);
    else if (req.intervalType === "monthly") sessionsPerWeek = Math.ceil(req.requiredMinutes / (SLOT_DURATION * 4));
    else if (req.intervalType === "daily") sessionsPerWeek = 5;

    let scheduled = 0;
    for (const day of days) {
      if (scheduled >= sessionsPerWeek) break;
      for (const time of timeSlots) {
        if (scheduled >= sessionsPerWeek) break;
        if (isSlotAvailable(provider.id, day, time)) {
          bookSlot(provider.id, day, time);
          proposedBlocks.push({
            id: -(proposedBlocks.length + 1), // temp negative id for proposed
            staffId: provider.id,
            staffName: `${provider.firstName} ${provider.lastName}`,
            studentId: req.studentId,
            studentName: req.studentFirst ? `${req.studentFirst} ${req.studentLast}` : null,
            serviceTypeId: req.serviceTypeId,
            serviceTypeName: req.serviceTypeName,
            dayOfWeek: day,
            startTime: time,
            endTime: addMinutes(time, SLOT_DURATION),
            location: null,
            blockLabel: req.serviceTypeName ?? null,
            blockType: "service",
            notes: `Auto-generated for week of ${weekOf}`,
            isRecurring: false,
            weekOf,
            isAutoGenerated: true,
            createdAt: new Date().toISOString(),
          });
          scheduled++;
        }
      }
    }

    if (scheduled < sessionsPerWeek) {
      unresolvedDeficits.push({
        studentId: req.studentId,
        studentName: req.studentFirst ? `${req.studentFirst} ${req.studentLast}` : `Student ${req.studentId}`,
        serviceRequirementId: req.id,
        serviceTypeName: req.serviceTypeName ?? "Unknown",
        gapDescription: `Could only schedule ${scheduled} of ${sessionsPerWeek} sessions for ${req.serviceTypeName}`,
        severity: scheduled === 0 ? "high" : "medium",
      });
    }
  }

  // Compute projected fulfillment
  const projectedFulfillment = await computeAllActiveMinuteProgress();

  res.json({
    weekOf,
    proposedBlocks,
    unresolvedDeficits,
    conflicts: [],
    projectedFulfillment,
    summary: `Generated ${proposedBlocks.length} proposed schedule blocks for week of ${weekOf}. ${unresolvedDeficits.length} unresolved gaps remain.`,
  });
});

router.post("/scheduler/accept", async (req, res): Promise<void> => {
  const parsed = AcceptGeneratedScheduleBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  // In a real system we'd store proposed blocks in a temp table and promote them
  // For now, we just return a success response
  res.json({ acceptedCount: parsed.data.blockIds.length, message: `Accepted ${parsed.data.blockIds.length} schedule blocks for week of ${parsed.data.weekOf}` });
});

// Staff Assignments
router.get("/staff-assignments", async (req, res): Promise<void> => {
  const params = ListStaffAssignmentsQueryParams.safeParse(req.query);
  const conditions: any[] = [];
  if (params.success) {
    if (params.data.staffId) conditions.push(eq(staffAssignmentsTable.staffId, Number(params.data.staffId)));
    if (params.data.studentId) conditions.push(eq(staffAssignmentsTable.studentId, Number(params.data.studentId)));
  }

  const assignments = await db
    .select({
      id: staffAssignmentsTable.id,
      staffId: staffAssignmentsTable.staffId,
      studentId: staffAssignmentsTable.studentId,
      assignmentType: staffAssignmentsTable.assignmentType,
      startDate: staffAssignmentsTable.startDate,
      endDate: staffAssignmentsTable.endDate,
      notes: staffAssignmentsTable.notes,
      createdAt: staffAssignmentsTable.createdAt,
      staffFirst: staffTable.firstName,
      staffLast: staffTable.lastName,
      staffRole: staffTable.role,
      studentFirst: studentsTable.firstName,
      studentLast: studentsTable.lastName,
    })
    .from(staffAssignmentsTable)
    .leftJoin(staffTable, eq(staffTable.id, staffAssignmentsTable.staffId))
    .leftJoin(studentsTable, eq(studentsTable.id, staffAssignmentsTable.studentId))
    .where(conditions.length > 0 ? and(...conditions) : undefined);

  res.json(assignments.map(a => ({
    ...a,
    staffName: a.staffFirst ? `${a.staffFirst} ${a.staffLast}` : null,
    staffRole: a.staffRole,
    studentName: a.studentFirst ? `${a.studentFirst} ${a.studentLast}` : null,
    createdAt: a.createdAt.toISOString(),
  })));
});

router.post("/staff-assignments", async (req, res): Promise<void> => {
  const parsed = CreateStaffAssignmentBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [assignment] = await db.insert(staffAssignmentsTable).values(parsed.data).returning();
  res.status(201).json({ ...assignment, createdAt: assignment.createdAt.toISOString() });
});

router.delete("/staff-assignments/:id", async (req, res): Promise<void> => {
  const params = DeleteStaffAssignmentParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  await db.delete(staffAssignmentsTable).where(eq(staffAssignmentsTable.id, params.data.id));
  res.sendStatus(204);
});

export default router;