# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MinuteOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedule-blocks:
    get:
      operationId: listScheduleBlocks
      tags: [schedules]
      summary: List schedule blocks
      parameters:
        - name: staffId
          in: query
          schema:
            type: ["integer", "null"]
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: weekOf
          in: query
          schema:
            type: ["string", "null"]
        - name: dayOfWeek
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Schedule blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleBlock"
    post:
      operationId: createScheduleBlock
      tags: [schedules]
      summary: Create schedule block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBlockBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleBlock"
  /schedule-blocks/{id}:
    patch:
      operationId: updateScheduleBlock
      tags: [schedules]
      summary: Update schedule block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleBlockBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleBlock"
    delete:
      operationId: deleteScheduleBlock
      tags: [schedules]
      summary: Delete schedule block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /schedule-blocks/conflicts:
    get:
      operationId: getScheduleConflicts
      tags: [schedules]
      summary: Detect schedule conflicts
      parameters:
        - name: weekOf
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Conflicts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleConflict"
...
  /scheduler/generate:
    post:
      operationId: generateSchedule
      tags: [schedules]
      summary: Run auto-schedule generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScheduleBody"
      responses:
        "200":
          description: Proposed schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedSchedule"
  /scheduler/accept:
    post:
      operationId: acceptGeneratedSchedule
      tags: [schedules]
      summary: Accept and apply generated schedule blocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptScheduleBody"
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptScheduleResult"

  # ─── Alerts ──────────────────────────────────────────────────────
...
  /staff-assignments:
    get:
      operationId: listStaffAssignments
      tags: [schedules]
      summary: List staff assignments
      parameters:
        - name: staffId
          in: query
          schema:
            type: ["integer", "null"]
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffAssignment"
    post:
      operationId: createStaffAssignment
      tags: [schedules]
      summary: Assign staff to student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffAssignmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAssignment"
  /staff-assignments/{id}:
    delete:
      operationId: deleteStaffAssignment
      tags: [schedules]
      summary: Remove staff assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Missed Reasons ───────────────────────────────────────────────
...
components:
  schemas:
...
    MinuteProgress:
      type: object
      properties:
        serviceRequirementId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: string
        providerId:
          type: ["integer", "null"]
        providerName:
          type: ["string", "null"]
        intervalType:
          type: string
        requiredMinutes:
          type: integer
        deliveredMinutes:
          type: integer
        remainingMinutes:
          type: integer
        percentComplete:
          type: number
        expectedMinutesByNow:
          type: number
        projectedMinutes:
          type: number
        riskStatus:
          type: string
        intervalStart:
          type: string
        intervalEnd:
          type: string
        missedSessionsCount:
          type: integer
        makeupSessionsCount:
          type: integer
      required: [serviceRequirementId, studentId, studentName, serviceTypeId, serviceTypeName, intervalType, requiredMinutes, deliveredMinutes, remainingMinutes, percentComplete, expectedMinutesByNow, projectedMinutes, riskStatus, intervalStart, intervalEnd, missedSessionsCount, makeupSessionsCount]

    # ─── Schedule Blocks ─────────────────────────────────────────
    ScheduleBlock:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        studentId:
          type: ["integer", "null"]
        studentName:
          type: ["string", "null"]
        serviceTypeId:
          type: ["integer", "null"]
        serviceTypeName:
          type: ["string", "null"]
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: ["string", "null"]
        blockLabel:
          type: ["string", "null"]
        blockType:
          type: string
        notes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        weekOf:
          type: ["string", "null"]
        isAutoGenerated:
          type: boolean
        createdAt:
          type: string
      required: [id, staffId, dayOfWeek, startTime, endTime, blockType, isRecurring, isAutoGenerated, createdAt]
    CreateScheduleBlockBody:
      type: object
      properties:
        staffId:
          type: integer
        studentId:
          type: ["integer", "null"]
        serviceTypeId:
          type: ["integer", "null"]
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: ["string", "null"]
        blockLabel:
          type: ["string", "null"]
        blockType:
          type: string
        notes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        weekOf:
          type: ["string", "null"]
      required: [staffId, dayOfWeek, startTime, endTime, blockType, isRecurring]
    UpdateScheduleBlockBody:
      type: object
      properties:
        studentId:
          type: ["integer", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ScheduleConflict:
      type: object
      properties:
        staffId:
          type: integer
        staffName:
          type: string
        dayOfWeek:
          type: string
        conflictingBlocks:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              startTime:
                type: string
              endTime:
                type: string
              studentName:
                type: ["string", "null"]
            required: [id, startTime, endTime]
        overlapMinutes:
          type: integer
      required: [staffId, staffName, dayOfWeek, conflictingBlocks, overlapMinutes]
    CoverageGap:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        serviceRequirementId:
          type: integer
        serviceTypeName:
          type: string
        gapDescription:
          type: string
        severity:
          type: string
      required: [studentId, studentName, serviceRequirementId, serviceTypeName, gapDescription, severity]

    # ─── Auto-scheduler ──────────────────────────────────────────
    GenerateScheduleBody:
      type: object
      properties:
        weekOf:
          type: string
        staffIds:
          type: array
          items:
            type: integer
        studentIds:
          type: array
          items:
            type: integer
        overwriteExisting:
          type: boolean
      required: [weekOf]
    GeneratedSchedule:
      type: object
      properties:
        weekOf:
          type: string
        proposedBlocks:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleBlock"
        unresolvedDeficits:
          type: array
          items:
            $ref: "#/components/schemas/CoverageGap"
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleConflict"
        projectedFulfillment:
          type: array
          items:
            $ref: "#/components/schemas/MinuteProgress"
        summary:
          type: string
      required: [weekOf, proposedBlocks, unresolvedDeficits, conflicts, projectedFulfillment, summary]
    AcceptScheduleBody:
      type: object
      properties:
        weekOf:
          type: string
        blockIds:
          type: array
          items:
            type: integer
      required: [weekOf, blockIds]
    AcceptScheduleResult:
      type: object
      properties:
        acceptedCount:
          type: integer
        message:
          type: string
      required: [acceptedCount, message]

    # ─── Alerts ──────────────────────────────────────────────────
...
    StaffAssignment:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        staffRole:
          type: ["string", "null"]
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        assignmentType:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, staffId, studentId, assignmentType, createdAt]
    CreateStaffAssignmentBody:
      type: object
      properties:
        staffId:
          type: integer
        studentId:
          type: integer
        assignmentType:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [staffId, studentId, assignmentType]

    # ─── Missed Reasons ──────────────────────────────────────────
...