# lib/api-zod/src/generated/types/coverageGap.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinuteOps API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoverageGap {
  studentId: number;
  studentName: string;
  serviceRequirementId: number;
  serviceTypeName: string;
  gapDescription: string;
  severity: string;
}