# lib/api-zod/src/generated/types/createStaffAssignmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinuteOps API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStaffAssignmentBody {
  staffId: number;
  studentId: number;
  assignmentType: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  notes?: string | null;
}